Revert "rename commits to logs to make space for conflicts"
This reverts commit dcf2f8cbbf4220c6e9b810ebbde7229235432166.
