+++ program/commits/view.go
@@ -1,4 +1,4 @@
-package logs
+package commits
 
 func (m Model) View() string {
 	return m.list.View()
